requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.1.0=gpu
absl-py=0.13.0=py38h06a4308_0
aiohttp=3.7.4=py38h27cfd23_1
astor=0.8.1=py38h06a4308_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.5=py38h261ae71_0
chardet=3.0.4=py38h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
configparser=5.0.2=pypi_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cython=0.29.23=py38h2531618_0
decorator=5.0.9=pypi_0
docker-pycreds=0.4.0=pypi_0
flatbuffers=1.12=pypi_0
gast=0.3.3=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.24=pypi_0
google-auth=1.32.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=py_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=py38hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
ipython=7.24.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.18.0=pypi_0
joblib=1.0.1=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
keras-tcn=3.4.0=pypi_0
keras-tuner=1.0.3=pypi_0
kt-legacy=1.0.3=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.3.0=hd4cf53a_17
logzero=1.7.0=pypi_0
markdown=3.3.4=py38h06a4308_0
matplotlib-inline=0.1.2=pypi_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
multidict=5.1.0=py38h27cfd23_2
ncurses=6.2=he6710b0_1
numpy=1.19.5=pypi_0
numpy-base=1.20.2=py38hfae3a4d_0
oauthlib=3.1.0=py_0
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=20.9=pypi_0
pandas=1.2.5=py38h295c915_0
parso=0.8.2=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pip=21.1.2=py38h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.14.0=py38h2531618_1
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.9.0=pypi_0
pyjwt=1.7.1=py38_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.10=h12debd9_8
python-dateutil=2.8.1=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.2=pypi_0
scipy=1.6.2=py38had2a1c9_1
sentry-sdk=1.4.3=pypi_0
setuptools=52.0.0=py38h06a4308_0
shortuuid=1.0.1=pypi_0
six=1.15.0=pypi_0
smmap=4.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=gpu_py38h8a7d6ce_0
tensorflow-addons=0.13.0=pypi_0
tensorflow-base=2.4.1=gpu_py38h29c2da4_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-gpu=2.4.1=h30adc30_0
termcolor=1.1.0=py38h06a4308_1
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
traitlets=5.0.5=pypi_0
typeguard=2.12.1=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
wandb=0.12.4=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py38h7b6447c_1
xz=5.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
yaspin=2.1.0=pypi_0
zerologger=1.1.4=pypi_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
